Fix and improve `txsub` package
This commit fixes an issue that caused sending `Timeout` responses to
client submitting transactions. The problem was present in
`TransactionByHashAfterLedger` function: it was checking the result of
the transaction in History DB and, when it was not found, in Core DB but
with ledger sequence after the latest ingested ledger. Given the fact
that Horizon ingests only successful transactions it was impossible to
find results for some of the failed transactions (especially those
failed at the consensus level that require more time to process).

`SubmissionTimeout` was changed to 30 seconds because HTTP clients in
SDKs usually timeout in 60 seconds. We want `SubmissionTimeout` to be
lower than that to make sure that they read the response before the
client timeout and 30 seconds is 6 ledgers (with avg. close time = 5
sec), enough for stellar-core to drop the transaction if not added to
the ledger and ask client to try again by sending a `Timeout` response.

It also contains improved logging in `txsub` package.